requirements.txt
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
Django==6.0.3
django-cors-headers==4.9.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
soupsieve==2.8.3
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
